Refactor template parameter handling into Parameters class
Introduced a new Parameters class to encapsulate template parameter logic and refactored Template to use this class. This improves code organization and maintainability by separating parameter management from the Template class.
